Remove theme submodule
The 'hugo-boxlist' theme was maintained as a submodule under the themes
folder.  It was a bit of a hassle to manage it as a submodule, and it's
probably not likely that the theme will be reused.  So we will add the
contents of the theme to the Xmeyers repo, rename the theme to
'boxlist'.  In the future if it looks like the theme can be reused, it
can be pulled out, and reincorporated as a git submodule.